tdf#83300 DOCX import: 'TOC Heading' should not be 'keep with next' by default
On one hand, neither our binary DOC import, nor Word maps the "TOC
Heading" style to something special, and that's how the DOCX import
added that property to some paragraphs in the document, moving the
as-char picture from the first to the second page.

OTOH, the DOCX export filter has a lcl_guessQFormat() function that
explicitly assumes that such a style name exists in Writer document
models, so again it doesn't make sense to handle this style name with
special care.

Change-Id: I3af548930f9683695fc3ad56b486e013f107d61a