qadevOOo: fix race in the ScAccessiblePreviewTable event test
XAccessibleEventBroadcaster::addAccessibleEventListener test calls the
fireEvent() and the problem is that the event is triggered not on the
print preview itself, but some toolbar button.  This then causes an
accessible event to be broadcast from the print preview too, but only
after doAccessibleAction() returns, some time later from VCL main-loop.

If the test checks the flag in the listener before the main thread sends
the event, the test fails; try to fix that with
XToolkit::processEventsToIdle().

Change-Id: Ia52bdf99cdc349ffb0a03536fba271b792caae0f