tdf#95888 sw: make click to header area show header controls again
Regression from commit 5cd01aa7dd6634286b9d26ef74d6e70085891a8f
(tdf#39080 Interactive hide-whitespace mode, 2015-08-30), the problem
was that SwPageFrm::IsOverHeaderFooterArea() checked if the header is
active before accepting that the user clicked in that area.

Don't require an active header at least in non-hide-whitespace mode, as
the member function is used in the "could there be a header here" sense
in SwEditWin::MouseButtonDown() to show the controls that allows
actually adding a header.

Change-Id: I6f905920113aed1512e333e718a8f26d88a3245d