vcl: reorder loop to avoid DrMemory warning
DrMemory 1.9.0-4 32-bit complains about UNINITIALIZED READ on the
instruction "test %ecx %ecx" corresponding to the *pW test in the while
loop, on the last iteration.

GetKeyNameTextW() is documented to null-terminate the buffer, so either
it has a bug and doesn't do that, or DrMemory has a bug and falsely
reports an error.

Either way it can be avoided by checking the end position first.

https://github.com/DynamoRIO/drmemory/issues/1819

Change-Id: I546d2057d39865d527c1d7c436f690669b214d75