sw lok comments: fix map mode state after changing zoom levels
Zoom level of SwEditWin is kept in sync with the client, so that the
pixel-based comment widgets can be positioned correctly. But that does
not mean in general the SwEditWin map mode should not be disabled: so
that we don't have to tweak the map mode for each and every
postMouseEvent() call and still be able to send them using logic
coordinates.

Change-Id: I6f686b93d2509d52fdd34e84a502cf04e1ce6e59