Non-implicit virtual SvxPageItem dtor
...otherwise, as the class is SVX_DLLPUBLIC and---for better or worse---
SVX_DLLPUBLIC is shared between Library_svxcore and Library_svx, all .cxx in
Library_svxcore that happen to include svx/pageitem.hxx would emit it (as well
as those in Library_svx, of course).  But at least clang-cl at /O0 would want
to put SvxPageItem's (base-class) vtable in place in the dtor, so would also
emit the vtable, so would depend on SvxPageItem::operator==, but that is defined
in pageitem.cxx in Library_svx, against which Library_svxcore doesn't link.

Change-Id: Ibfc393c1de3e0bd6621c057d88b1b0f272820154