tdf#96234: Type Formatting, Primary key checkbox may be checked
Right click on a field in Type Formatting (Base, paste a table from Calc for eg)
makes a popup which takes into account now if the field is a primary key or not.

Handles multi-fields primary key, too: the selected fields must be exactly
those in the primary key, no more, no less.

Change-Id: I545b946bccd3fc02f4f6c3bdb99f914d181a3d40