ucb: webdav-curl: tdf#95792: fix saving file the first time on some WebDAV servers.
Some WebDAV servers don't implement section 7.3 of RFC4918:
<http://tools.ietf.org/html/rfc4918#section-7.3>

This lack of implementation breaks 'Save As...' functionality
when the target is a WebDAV server, by not locking the URL
('reserve the name for use', in RFC4918 parlance).

The server not implementing this usually answers with one of
'405 Method Not Allowed', '501 Not Implemented' or
'412 Precondition Failed' http error codes.

The fix should manage this lack of implementation.

[ port of commit 6255f64ffe8e79a4aede63c71936d5857ff6822b ]

Change-Id: I343368bb91bdffff143aa26d8449b48f3b5300d8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123275
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>