sticky end col/row anchor for range references, tdf#92779
In range references referring more than one column/row where the end
col/row points to the maximum column or row number that col/row is not
decremented when the range is shrunken. Incrementing an end col/row is
not done past the maximum column or row number, so such references do
not yield #REF! errors anymore. This is also done in named expressions
if the end col/row is an absolute reference.

Change-Id: Iafa2d62abd3e816a1c56c3166af92807e55b75ce