tdf#92925 Remove MET (OS/2) export
MET Files were used with OS/2 and are not used anymore.
I did test import which this patch did not remove.  However,
with MET files you can not export them as an image using "Save as
image" (this patch didn't cause or fix that).  You can still use
the export of the whole document to convert the MET file.

Change-Id: Ie818ef963db14166be354379d4b2991a3b72b645
Reviewed-on: https://gerrit.libreoffice.org/20622
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>