tdf#96119 Cannot select item with mouse in any combobox / drop down menu
this commit has make the problem, the removing of the SALEVENT_MOUSEACTIVATE

commit 813fe9a98b9093feeb3025dfed87fc94dcb8199e
Author: Noel Grandin <noel@peralex.com>
Date:   Tue Nov 24 08:50:39 2015 +0200

    loplugin:unusedfields in vcl/

    and remove the unused SALEVENT_MOUSEACTIVATE stuff

Without of there, when click in the opened list-box it send first the
"PreNotify" with "MouseNotifyEvent::LOSEFOCUS" and that close the listbox.
After that, it send the mouse-event to a closed window, that is the reason
why the Listbox not get the mouse-click.

With this patch, first send the mouse-click and then the "PreNotify"

Change-Id: I5a09b1524335434f043d22bc71f7e38559fb1c0b
Reviewed-on: https://gerrit.libreoffice.org/20708
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>