Use #version 140 to get inverse()
The Linux GLSL imlementation is very loose in its interpretation of
standards, it seems. (I am not surprised.) Not so on Windows, at least
with AMD drivers. The inverse() function (for float matririces) is
present only starting from GLSL 1.40. Bump the #version directive
accordingly. (See
https://www.opengl.org/sdk/docs/man/html/inverse.xhtml)

To get this to work again on OS X will be a further challenge. There a
'legacy' context only has GLSL 1.20. But the intent is to move this
code in the direction of using 'core' contexts anyway.

Change-Id: I572dc7cc30f90321b233fbfd857455dc01761a7c