The likely useless ImplDoTiledRendering() is used in vcl only
So move it to a locale include file in vcl. Now <vcl/window.hxx> does
not need to include <config_features.h>.

As such it is questionable whether what ImplDoTiledRendering()
currently does makes any sense; de facto it now just means that we are
on iOS. We have comphelper::LibreOfficeKit::isActive() which probably
is what should be used to test whether tiled rendering is used. At
least in the cases where LibreOfficeKit's "unstable" API is used for
tasks that involve rendering of tiles. Hmm, that sounds like a
circular argument.

Change-Id: I52c984aec7510526c5294579966345b8340c8642