tdf#96366 disable treeview results until fix available
Since 66944ae2897c46e4a5ad8e6ed1eaf284bf4e9be4 EditThisFunc() and
EditNextFunc() are used to iterate through the formula to obtain
expression results to display in the treeview. Calling the Edit...()
functions spoils about everything as it messes around with the edit
state. As the name suggests..

Change-Id: I8b35d85b7bbf8821b5a995e84f9dd88a0f6f00b9