tdf#96713 OdfFlatXml: Seek to 0 before reading
Similar to tdf#45418.

The problem is that sfx2 DocumentInserter code calls SfxMedium::IsStorage,
which reads the stream but doesn't correctly seek back to 0. Actually
SfxMedium_Impl has 2 members for the input stream, one of SvStream type
and another one as Reference<XInputStream>. Turns out that reading with
SvStream::Read changes the position in the object referenced by
Reference<XInputStream>, but SvStream::Seek doesn't, so Seek(0) doesn't
do the desired effect.

My current solution is to ensure that we're reading from 0 inside the
filter. I think it's a good thing to do anyway, and should be sufficient,
given that other filters doesn't seem to be affected by this bug.

Change-Id: I49b41077032d3e395c675e4f7824cc778c075473