tdf#75191: only call atk_register_role if strictly necessary
As part of tdf#39944, the mappings of UNO and ATK roles have been
reviewed and right now there is a match in ATK for every UNO role.

Still, some protection must be done in case of running LO with an
old ATK version that does not contain all the roles we need. The
function registerRole serves this purpose now, and it has been
renamed consequently.

Finally, removed a redundant mapping of the HEADING role.

Change-Id: I7a19c3f3efa5d6ff2697c19e3ae7856d4702a253