Revert "vcl: move filter headers to vcl/inc/filter"
Discussed on IRC - overwhelming consensus is to keep the headers
where they were before. The main reason to keep them in the same
directory is to keep headers as local as possible to the source
files, unless something else needs to access it.

The only thing that uses the headers is vc/workben/fftester.cxx

I will work out how to get Doxygen to pick up these headers another way.

This reverts commit 2054cdfd9bbfead3da63c4924a781924dba8f10e.

Change-Id: I45ee970722e27748ab8af8909ac49579dc47ef56
Reviewed-on: https://gerrit.libreoffice.org/21250
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>
Tested-by: Chris Sherlock <chris.sherlock79@gmail.com>