vcl: Rename class ImplFTSFontData to FreetypeFontFace
ImplFTSFontData derives from PhysicalFontFace, and implements the
mechanisms specified therein. It's not actually a "data" class, but
also implements a factory function that creates a
LogicalFontInstance object via Freetype. I believe the "S" in "FTS"
stands for "Server", which is a misnomer as Freetype is a font
renderer, not a font server.

Change-Id: Ibacc4bf3f850ddfb37016ea2e3e088279b1b1ba1
Reviewed-on: https://gerrit.libreoffice.org/21285
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>